+++ src/vs/workbench/services/search/test/browser/queryBuilder.test.ts
@@ -6,7 +6,7 @@ import * as assert from 'assert';
 import { IExpression } from 'vs/base/common/glob';
 import { join } from 'vs/base/common/path';
 import { isWindows } from 'vs/base/common/platform';
-import { URI as uri } from 'vs/base/common/uri';
+import { URI, URI as uri } from 'vs/base/common/uri';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
 import { TestConfigurationService } from 'vs/platform/configuration/test/common/testConfigurationService';
 import { TestInstantiationService } from 'vs/platform/instantiation/test/common/instantiationServiceMock';
@@ -551,7 +551,7 @@ suite('QueryBuilder', () => {
 			});
 	});
 
-	suite('parseSearchPaths', () => {
+	suite('parseSearchPaths 1', () => {
 		test('simple includes', () => {
 			function testSimpleIncludes(includePattern: string, expectedPatterns: string[]): void {
 				const result = queryBuilder.parseSearchPaths(includePattern);
@@ -920,6 +920,46 @@ suite('QueryBuilder', () => {
 							}]
 					}
 				]
+			];
+			cases.forEach(testIncludesDataItem);
+		});
+	});
+
+	suite('parseSearchPaths 2', () => {
+
+		function testIncludes(includePattern: string, expectedResult: ISearchPathsInfo): void {
+			assertEqualSearchPathResults(
+				queryBuilder.parseSearchPaths(includePattern),
+				expectedResult,
+				includePattern);
+		}
+
+		function testIncludesDataItem([includePattern, expectedResult]: [string, ISearchPathsInfo]): void {
+			testIncludes(includePattern, expectedResult);
+		}
+
+		test('includes with tilde', () => {
+			const userHome = URI.file('/');
+			const cases: [string, ISearchPathsInfo][] = [
+				[
+					'~/foo/bar',
+					{
+						searchPaths: [{ searchPath: getUri(userHome.fsPath, '/foo/bar') }]
+					}
+				],
+				[
+					'~/foo/bar, a',
+					{
+						searchPaths: [{ searchPath: getUri(userHome.fsPath, '/foo/bar') }],
+						pattern: patternsToIExpression(...globalGlob('a'))
+					}
+				],
+				[
+					fixPath('/foo/~/bar'),
+					{
+						searchPaths: [{ searchPath: getUri('/foo/~/bar') }]
+					}
+				],
 			];
 			cases.forEach(testIncludesDataItem);
 		});
